silver tests refac
