+++ source/core/tests/helpers/schemas/bronze_calculated_measurements_schema.py
