+++ source/core/tests/helpers/schemas/bronze_submitted_transactions_value_schema.py
