+++ source/core/tests/helpers/schemas/silver_measurements_schema.py
