+++ source/core/tests/integration/silver/infrastructure/repositories/test_silver_measurements_repository.py
@@ -14,28 +14,6 @@ from core.silver.infrastructure.repositories.silver_measurements_repository impo
 from tests.helpers.builders.silver_measurements_builder import SilverMeasurementsBuilder
 
 
-def test__write_stream__called__with_correct_arguments(mock_checkpoint_path: mock.MagicMock | mock.AsyncMock) -> None:
-    # Arrange
-    mocked_measurements = mock.Mock()
-    mocked_batch_operation = mock.Mock()
-
-    # Act
-    SilverMeasurementsRepository().write_stream(
-        mocked_measurements,
-        GehCommonOrchestrationType.SUBMITTED,
-        mocked_batch_operation,
-    )
-
-    # Assert
-    mocked_measurements.writeStream.outputMode.assert_called_once_with("append")
-    mocked_measurements.writeStream.outputMode().format().option().trigger.assert_called_once()
-    mocked_measurements.writeStream.outputMode().format().option().trigger().foreachBatch.assert_called_once_with(
-        mocked_batch_operation
-    )
-    mocked_measurements.writeStream.outputMode().format().option().trigger().foreachBatch().start.assert_called_once()
-    mocked_measurements.writeStream.outputMode().format().option().trigger().foreachBatch().start().awaitTermination.assert_called_once()
-
-
 def test__write_measurements__when_contionous_streaming_is_disabled__should_not_call_trigger() -> None:
     # Arrange
     os.environ["CONTINUOUS_STREAMING_ENABLED"] = "true"
