+++ source/core/tests/integration/silver/infrastructure/repositories/test_submitted_transactions_repository.py
@@ -1,7 +1,4 @@
-from unittest.mock import Mock
-
 from geh_common.domain.types.orchestration_type import OrchestrationType
-from geh_common.domain.types.orchestration_type import OrchestrationType as GehCommonOrchestrationType
 from pyspark.sql import SparkSession
 
 import tests.helpers.identifier_helper as identifier_helper
@@ -11,23 +8,6 @@ from core.silver.domain.constants.column_names.silver_measurements_column_names 
 from core.silver.infrastructure.config import SilverTableNames
 from core.silver.infrastructure.repositories.submitted_transactions_repository import SubmittedTransactionsRepository
 from tests.helpers.builders.silver_measurements_builder import SilverMeasurementsBuilder
-
-
-def test__read_submitted_transactions__calls_expected() -> None:
-    # Arrange
-    mock_spark = Mock()
-    expected_table = f"{SilverSettings().silver_database_name}.{SilverTableNames.silver_measurements}"
-
-    # Act
-    _ = SubmittedTransactionsRepository(mock_spark).read()
-
-    # Assert
-    mock_spark.readStream.format.assert_called_once_with("delta")
-    mock_spark.readStream.format().option.assert_called_once_with("ignoreDeletes", "true")
-    mock_spark.readStream.format().option().table.assert_called_once_with(expected_table)
-    mock_spark.readStream.format().option().table().filter.assert_called_once_with(
-        f"{SilverMeasurementsColumnNames.orchestration_type} = '{GehCommonOrchestrationType.SUBMITTED.value}'"
-    )
 
 
 def test__read_submitted_transaction__returns_expected(spark: SparkSession, migrations_executed) -> None:
