--- source/core/tests/silver/.gitignore
@@ -1,5 +1,0 @@
-*.local.settings.yml
-
-# Exclude folders generated by tests
-**/__*__
-metastore_db
