+++ source/core/tests/unit_tests/silver/application/streams/test_submitted_transactions.py
@@ -4,7 +4,6 @@ from pytest_mock import MockerFixture
 import tests.helpers.identifier_helper as identifier_helper
 import tests.helpers.table_helper as table_helper
 from core.bronze.infrastructure.config.table_names import TableNames as BronzeTableNames
-from core.contracts.process_manager.enums.metering_point_type import MeteringPointType
 from core.settings.bronze_settings import BronzeSettings
 from core.settings.silver_settings import SilverSettings
 from core.silver.application.streams import submitted_transactions as sut
@@ -52,34 +51,3 @@ def test__submitted_transactions__should_save_in_silver_measurements(
         f"orchestration_instance_id = '{expected_orchestration_id}'"
     )
     assert silver_table.count() == 1
-
-
-def test__stream_submitted_transactions__when_invalid_should_save_in_bronze_submitted_transactions_quarantined(
-    mock_checkpoint_path, spark: SparkSession, migrations_executed
-) -> None:
-    # Arrange
-    bronze_settings = BronzeSettings()
-    expected_orchestration_id = identifier_helper.generate_random_string()
-    value = (
-        ValueBuilder(spark)
-        .add_row(
-            orchestration_instance_id=expected_orchestration_id,
-            metering_point_type=MeteringPointType.MPT_UNSPECIFIED.value,
-        )
-        .build()
-    )
-    submitted_transactions = SubmittedTransactionsBuilder(spark).add_row(value=value).build()
-    table_helper.append_to_table(
-        submitted_transactions,
-        bronze_settings.bronze_database_name,
-        BronzeTableNames.bronze_submitted_transactions_table,
-    )
-
-    # Act
-    sut.stream_submitted_transactions()
-
-    # Assert
-    bronze_submitted_transactions_quarantined_table = spark.table(
-        f"{bronze_settings.bronze_database_name}.{BronzeTableNames.bronze_submitted_transactions_quarantined}"
-    ).where(f"orchestration_instance_id = '{expected_orchestration_id}'")
-    assert bronze_submitted_transactions_quarantined_table.count() == 1
