+++ source/core/tests/unit_tests/silver/domain/validations/test_enum_validations.py
