+++ source/core/tests/unit_tests/silver/infrastructure/protobuf/test_persist_submitted_transaction.py
